refactor: 이미지 관련 util 위치 수정
