+++ src/main/java/com/youthexpedition/azit/modules/crew/application/service/mapper/CrewResponseMapper.java
@@ -1,6 +1,6 @@
 package com.youthexpedition.azit.modules.crew.application.service.mapper;
 
-import com.youthexpedition.azit.infrastructure.common.util.ImageUrlFormatUtil;
+import com.youthexpedition.azit.infrastructure.common.util.image.ImageUrlFormatUtil;
 import com.youthexpedition.azit.modules.crew.application.port.in.dto.CreateCrewResponse;
 import com.youthexpedition.azit.modules.crew.application.port.in.dto.CrewInfoResponse;
 import com.youthexpedition.azit.modules.crew.application.port.in.dto.CrewInvitationResponse;
