+++ src/main/java/com/youthexpedition/azit/modules/member/application/service/mapper/MemberResponseMapper.java
@@ -1,6 +1,6 @@
 package com.youthexpedition.azit.modules.member.application.service.mapper;
 
-import com.youthexpedition.azit.infrastructure.common.util.ImageUrlFormatUtil;
+import com.youthexpedition.azit.infrastructure.common.util.image.ImageUrlFormatUtil;
 import com.youthexpedition.azit.modules.crew.domain.model.Crew;
 import com.youthexpedition.azit.modules.crew.domain.model.CrewMember;
 import com.youthexpedition.azit.modules.crew.domain.model.enums.CrewMemberRole;
